+++ src/features/breadcrumbs/components/Ellipsis/Ellipsis.module.css
